[#116] refactor: 포스트 쿼리 패키지 구조 수정
