+++ src/main/java/com/mallang/post/query/PostQueryService.java
@@ -1,8 +1,13 @@
 package com.mallang.post.query;
 
-import com.mallang.post.query.dao.PostDetailDao;
-import com.mallang.post.query.dao.PostSearchDao;
-import com.mallang.post.query.dao.PostSearchDao.PostSearchCond;
+import com.mallang.auth.domain.Member;
+import com.mallang.auth.domain.MemberRepository;
+import com.mallang.post.domain.Post;
+import com.mallang.post.domain.PostVisibilityPolicy.Visibility;
+import com.mallang.post.exception.NoAuthorityAccessPostException;
+import com.mallang.post.query.repository.PostLikeQueryRepository;
+import com.mallang.post.query.repository.PostQueryRepository;
+import com.mallang.post.query.repository.PostSearchDao.PostSearchCond;
 import com.mallang.post.query.response.PostDetailResponse;
 import com.mallang.post.query.response.PostSearchResponse;
 import jakarta.annotation.Nullable;
@@ -17,23 +22,56 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class PostQueryService {
 
-    private final PostDetailDao postDetailDao;
-    private final PostSearchDao postSearchDao;
-    private final PostDataValidator postDataValidator;
-    private final PostDataProtector postDataProtector;
+    private final MemberRepository memberRepository;
+    private final PostQueryRepository postQueryRepository;
+    private final PostLikeQueryRepository postLikeQueryRepository;
 
-    public PostDetailResponse getByIdAndBlogName(Long id,
-                                                 String blogName,
-                                                 @Nullable Long memberId,
-                                                 @Nullable String postPassword) {
-        PostDetailResponse postDetailResponse = postDetailDao.find(id, blogName, memberId);
-        postDataValidator.validateAccessPost(memberId, postDetailResponse);
-        return postDataProtector.protectIfRequired(memberId, postPassword, postDetailResponse);
+    public PostDetailResponse getByIdAndBlogName(
+            Long id,
+            String blogName,
+            @Nullable Long memberId,
+            @Nullable String postPassword
+    ) {
+        Member member = findMember(memberId);
+        Post post = postQueryRepository.getByIdAndBlogName(id, blogName);
+        try {
+            post.validatePostAccessibility(member, postPassword);
+            if (memberId == null) {
+                return PostDetailResponse.from(post);
+            }
+            boolean isLiked = postLikeQueryRepository.existsByMemberIdAndPostId(memberId, id, blogName);
+            return PostDetailResponse.withLiked(post, isLiked);
+        } catch (NoAuthorityAccessPostException e) {
+            if (post.getVisibilityPolish().getVisibility() == Visibility.PRIVATE) {
+                throw e;
+            }
+            return PostDetailResponse.protectedPost(post);
+        }
     }
 
-    public Page<PostSearchResponse> search(PostSearchCond cond, Pageable pageable, @Nullable Long memberId) {
-        Page<PostSearchResponse> result = postSearchDao.search(memberId, cond, pageable);
-        return postDataProtector.protectIfRequired(memberId, result);
+    public Page<PostSearchResponse> search(
+            PostSearchCond cond,
+            Pageable pageable,
+            @Nullable Long memberId
+    ) {
+        Member member = findMember(memberId);
+        return postQueryRepository.search(memberId, cond, pageable)
+                .map(post -> {
+                    try {
+                        post.validatePostAccessibility(member, null);
+                        return PostSearchResponse.from(post);
+                    } catch (NoAuthorityAccessPostException e) {
+                        return PostSearchResponse.protectedPost(post);
+                    }
+                });
+    }
+
+    @Nullable
+    private Member findMember(@Nullable Long memberId) {
+        if (memberId == null) {
+            return null;
+        }
+        return memberRepository.getById(memberId);
     }
 }
