+++ src/main/java/com/mallang/post/query/PostStarQueryService.java
@@ -1,6 +1,10 @@
 package com.mallang.post.query;
 
-import com.mallang.post.query.dao.StaredPostDao;
+import com.mallang.auth.domain.Member;
+import com.mallang.auth.domain.MemberRepository;
+import com.mallang.post.domain.Post;
+import com.mallang.post.exception.NoAuthorityAccessPostException;
+import com.mallang.post.query.repository.PostStarQueryRepository;
 import com.mallang.post.query.response.StaredPostResponse;
 import jakarta.annotation.Nullable;
 import lombok.RequiredArgsConstructor;
@@ -14,16 +18,33 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class PostStarQueryService {
 
-    private final StaredPostDao staredPostDao;
-    private final PostDataProtector postDataProtector;
+    private final MemberRepository memberRepository;
+    private final PostStarQueryRepository postStarQueryRepository;
 
     public Page<StaredPostResponse> findAllByMemberId(
             Long targetMemberId,
             @Nullable Long requesterId,
             Pageable pageable
     ) {
-        Page<StaredPostResponse> staredPostData = staredPostDao.find(targetMemberId, pageable);
-        return postDataProtector.protectStaredIfRequired(requesterId, staredPostData);
+        Member member = findMember(requesterId);
+        return postStarQueryRepository.findAllByMemberId(targetMemberId, pageable)
+                .map(postStar -> {
+                    try {
+                        Post post = postStar.getPost();
+                        post.validatePostAccessibility(member, null);
+                        return StaredPostResponse.from(postStar);
+                    } catch (NoAuthorityAccessPostException e) {
+                        return StaredPostResponse.protectedPost(postStar);
+                    }
+                });
+    }
+
+    @Nullable
+    private Member findMember(@Nullable Long memberId) {
+        if (memberId == null) {
+            return null;
+        }
+        return memberRepository.getById(memberId);
     }
 }
