+++ src/main/java/com/mallang/post/query/repository/DraftQueryRepository.java
@@ -1,4 +1,4 @@
-package com.mallang.post.query.support;
+package com.mallang.post.query.repository;
 
 import com.mallang.blog.domain.Blog;
 import com.mallang.post.domain.draft.Draft;
@@ -6,7 +6,7 @@ import com.mallang.post.exception.NotFoundDraftException;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DraftQuerySupport extends JpaRepository<Draft, Long> {
+public interface DraftQueryRepository extends JpaRepository<Draft, Long> {
 
     default Draft getById(Long id) {
         return findById(id).orElseThrow(NotFoundDraftException::new);
