+++ src/main/java/com/mallang/post/query/repository/PostStarQueryRepository.java
@@ -1,11 +1,13 @@
-package com.mallang.post.query.support;
+package com.mallang.post.query.repository;
 
 import com.mallang.post.domain.star.PostStar;
 import java.util.List;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostStarQuerySupport extends JpaRepository<PostStar, Long> {
+public interface PostStarQueryRepository extends
+        JpaRepository<PostStar, Long>,
+        StaredPostDao {
 
     @EntityGraph(attributePaths = {"post", "member"})
     List<PostStar> findAllWithPostAndMemberByMemberIdOrderByIdDesc(Long memberId);
