+++ src/main/java/com/mallang/statistics/api/query/dao/PostTotalViewsDao.java
@@ -4,7 +4,7 @@ package com.mallang.statistics.api.query.dao;
 import static com.mallang.statistics.statistic.QPostViewStatistic.postViewStatistic;
 
 import com.mallang.post.domain.Post;
-import com.mallang.post.query.support.PostQuerySupport;
+import com.mallang.post.query.repository.PostQueryRepository;
 import com.mallang.statistics.api.query.response.PostTotalViewsResponse;
 import com.mallang.statistics.statistic.PostViewStatistic;
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -15,11 +15,11 @@ import org.springframework.stereotype.Component;
 @Component
 public class PostTotalViewsDao {
 
-    private final PostQuerySupport postQuerySupport;
+    private final PostQueryRepository postQueryRepository;
     private final JPAQueryFactory query;
 
     public PostTotalViewsResponse find(String blogName, Long postId) {
-        Post post = postQuerySupport.getByIdAndBlogName(postId, blogName);
+        Post post = postQueryRepository.getByIdAndBlogName(postId, blogName);
         int totalCount = query.selectFrom(postViewStatistic)
                 .where(postViewStatistic.postId.eq(post.getPostId()))
                 .fetch()
