+++ src/test/java/com/mallang/post/query/PostQueryServiceTest.java
@@ -12,7 +12,7 @@ import com.mallang.post.application.command.ClickPostLikeCommand;
 import com.mallang.post.application.command.CreatePostCommand;
 import com.mallang.post.exception.BadPostSearchCondException;
 import com.mallang.post.exception.NoAuthorityAccessPostException;
-import com.mallang.post.query.dao.PostSearchDao.PostSearchCond;
+import com.mallang.post.query.repository.PostSearchDao.PostSearchCond;
 import com.mallang.post.query.response.PostDetailResponse;
 import com.mallang.post.query.response.PostSearchResponse;
 import com.mallang.post.query.response.PostSearchResponse.CategoryResponse;
